src/django_app/requirements.txt
# Django and extensions
Django==5.0.1
django-environ==0.11.2
django-extensions==3.2.3
django-cors-headers==4.3.1
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9
redis==5.0.1

# Celery
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# ASGI/WSGI servers
gunicorn==21.2.0
uvicorn[standard]==0.25.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Monitoring and metrics
prometheus-client==0.19.0
django-prometheus==2.3.1

# Network tools
netmiko==4.3.0
paramiko==3.4.0
psutil==5.9.6

# Data processing
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0

# API and serialization
pydantic==2.5.3
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Development tools
ipython==8.18.1
black==23.12.1
flake8==7.0.0
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
aiohttp==3.9.1

# Security
django-ratelimit==4.1.0
django-filter==23.5